infra: alert() joins the other console-widget bindings in the UI file
alert() was the last QApplication user in TLuaInterpreter.cpp, and it is
a UI function: it flashes the main window to get the player's attention.
It belongs beside the other Lua bindings that drive the widget layer,
for the same reason the eight before it moved.

The count is unchanged at 149, and that is expected rather than a failed
ratchet. TLuaInterpreterUI.cpp was clean until now, so the offence moves
from one file to the other rather than disappearing: the audit counts
the QApplication symbol wherever it sits, whether or not the header is
included directly. What the move buys is where the offence lives. The
plan takes TLuaInterpreterUI.cpp wholesale to the future mudlet_app
target, so an offence in that file leaves core when the file does, while
the same offence in TLuaInterpreter.cpp would have to be dealt with on
its own later. None of the nine functions moved so far ratcheted the
count either.

QApplication::alert() is kept as it is. Reaching the same effect through
QWindow::alert() off the window handle would be a behaviour-adjacent
rewrite, which does not belong in a relocation.

The body is byte-identical and the lua_register() line stays put.
docs/libmudlet-widgets-report.md is regenerated: the two files swap
tables. cmake/core-widgets-baseline.txt is unchanged at 149.
