+++ src/TLuaInterpreterUI.cpp
@@ -29,6 +29,7 @@
 
 #include "TLuaInterpreter.h"
 
+#include <QApplication>
 #include <QClipboard>
 #include <QGuiApplication>
 
@@ -264,6 +265,25 @@ int TLuaInterpreter::addMouseEvent(lua_State* L)
 
     lua_pushboolean(L, true);
     return 1;
+}
+
+// Documentation: https://wiki.mudlet.org/w/Manual:Lua_Functions#alert
+int TLuaInterpreter::alert(lua_State* L)
+{
+    double luaAlertDuration = 0.0;
+
+    if (lua_gettop(L) > 0) {
+        luaAlertDuration = getVerifiedDouble(L, __func__, 1, "alert duration in seconds");
+        if (luaAlertDuration < 0.000) {
+            lua_pushstring(L, "alert: duration, in seconds, is optional but if given must be zero or greater.");
+            return lua_error(L);
+        }
+    }
+
+    // QApplication::alert expects milliseconds, not seconds
+    QApplication::alert(mudlet::self(), qRound(luaAlertDuration * 1000.0));
+
+    return 0;
 }
 
 // Documentation: https://wiki.mudlet.org/w/Manual:Lua_Functions#appendBuffer
